typing, launch to main.py
